refactor details section
